@@ -6,7 +6,16 @@
  * @package    Util
  * @subpackage UnitTests
  */
-class Horde_Util_DomhtmlTest extends PHPUnit_Framework_TestCase
+
+namespace Horde\Util\Test\Unnamespaced;
+
+use DOMElement;
+use DOMNode;
+use PHPUnit\Framework\TestCase;
+use Horde_Domhtml;
+use Horde_String;
+
+class DomhtmlTest extends TestCase
 {
     public function testBug9567()
     {
@@ -89,7 +98,7 @@ EOT;
         $dom = new Horde_Domhtml($text, 'iso-8859-2');
         $this->assertEquals(
             Horde_String::convertCharset($expected, 'UTF-8', 'iso-8859-2'),
-            strtr(trim($dom->returnBody()), array("\n" => ''))
+            strtr(trim($dom->returnBody()), ["\n" => ''])
         );
 
         $this->assertEquals(
@@ -100,18 +109,19 @@ EOT;
 
     public function testIterator()
     {
-        $text = file_get_contents(__DIR__ . '/fixtures/domhtml_test.html');
+        $text = file_get_contents(__DIR__ . '/../fixtures/domhtml_test.html');
         $dom = new Horde_Domhtml($text);
 
-        $tags = array(
+        $tags = [
             'html',
             'body',
             'div',
             'head',
             'title'
-        );
+        ];
 
         foreach ($dom as $node) {
+            $this->assertInstanceOf(DOMNode::class, $node);
             if ($node instanceof DOMElement) {
                 if ($node->tagName != reset($tags)) {
                     $this->fail('Wrong tag name.');
@@ -184,7 +194,7 @@ EOT;
 
         $this->assertEquals(
             $dom->returnHtml(),
-            $dom->returnHtml(array('charset' => 'iso-8859-1'))
+            $dom->returnHtml(['charset' => 'iso-8859-1'])
         );
     }
 
@@ -192,10 +202,9 @@ EOT;
     {
         $dom = new Horde_Domhtml('<html><body><div>foo</div></body></html>', 'UTF-8');
 
-        $this->assertRegExp(
+        $this->assertMatchesRegularExpression(
             '/"text\/html; charset=utf-8"/',
-            $dom->returnHtml(array('metacharset' => true))
+            $dom->returnHtml(['metacharset' => true])
         );
     }
-
 }
